refactor(apps/api_test): refactor apps/api_test, move to apis/web/api_test
