+++ src/dashboard/apigateway/apigateway/biz/permission.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -17,10 +16,23 @@
 # to the current version of the project delivered to anyone in the future.
 #
 
+from apigateway.apps.permission.constants import GrantTypeEnum
+from apigateway.apps.permission.models import AppResourcePermission
 
-def render_path(path, path_params):
-    for key, value in path_params.items():
-        tpl_key = "{%s}" % key
-        if path.find(tpl_key) >= 0:
-            path = path.replace(tpl_key, value)
-    return path
+
+class ResourcePermissionHandler:
+    @staticmethod
+    def grant(gateway, resource_id, bk_app_code, expire_days):
+        app_resource_permission = AppResourcePermission.objects.get_permission_or_none(
+            gateway=gateway,
+            resource_id=resource_id,
+            bk_app_code=bk_app_code,
+        )
+        if not app_resource_permission or app_resource_permission.will_expired_in(seconds=300):
+            AppResourcePermission.objects.save_permissions(
+                gateway=gateway,
+                resource_ids=[resource_id],
+                bk_app_code=bk_app_code,
+                grant_type=GrantTypeEnum.INITIALIZE.value,
+                expire_days=expire_days,
+            )
